+++ barretenberg/cpp/src/CMakeLists.txt
@@ -58,6 +58,7 @@ add_subdirectory(barretenberg/eccvm)
 add_subdirectory(barretenberg/env)
 add_subdirectory(barretenberg/examples)
 add_subdirectory(barretenberg/flavor)
+add_subdirectory(barretenberg/goblin)
 add_subdirectory(barretenberg/grumpkin_srs_gen)
 add_subdirectory(barretenberg/honk)
 add_subdirectory(barretenberg/join_split_example)
